Move Mock-LLM chat-navigation e2e cases to the many-models scenario
